requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jwt==1.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
